Remove core link to nextdom
