+++ assets/js/core/plugin.ajax.js
@@ -250,7 +250,7 @@ $('.eqLogicAction[data-action=copy]').on('click', function () {
 });
 
 $('.eqLogicAction[data-action=export]').on('click', function () {
-  window.open('core/php/export.php?type=eqLogic&id=' + $('.li_eqLogic.active').attr('data-eqLogic_id'), "_blank", null);
+  window.open('src/Api/export.php?type=eqLogic&id=' + $('.li_eqLogic.active').attr('data-eqLogic_id'), "_blank", null);
 });
 
 $('.eqLogicAttr[data-l1key=name]').on('change', function () {
