+++ assets/js/desktop/tools/backup.js
@@ -250,8 +250,9 @@ function initEvents() {
 function getNextDomLog(_autoUpdate, _log) {
     $.ajax({
         type: 'POST',
-        url: 'core/ajax/log.ajax.php',
+        url: 'src/ajax.php',
         data: {
+            target: 'Log',
             action: 'get',
             log: _log,
         },
