+++ src/Com/ComHttp.php
@@ -23,7 +23,7 @@ namespace NextDom\Com;
 use NextDom\Exceptions\CoreException;
 use NextDom\Helpers\LogHelper;
 
-require_once __DIR__ . '/../../core/php/core.inc.php';
+require_once __DIR__ . '/../../src/core.php';
 
 class ComHttp
 {
