+++ src/Com/ComShell.php
@@ -20,7 +20,7 @@ namespace NextDom\Com;
 
 use NextDom\Exceptions\CoreException;
 
-require_once __DIR__ . '/../../core/php/core.inc.php';
+require_once __DIR__ . '/../../src/core.php';
 
 class ComShell
 {
