+++ src/Helpers/TimeLineHelper.php
@@ -34,6 +34,7 @@
 
 namespace NextDom\Helpers;
 
+use NextDom\Com\ComShell;
 use NextDom\Managers\ConfigManager;
 
 /**
@@ -64,7 +65,7 @@ class TimeLineHelper
         if (!file_exists($path)) {
             $result = [];
         } else {
-            \com_shell::execute(SystemHelper::getCmdSudo() . 'chmod 666 ' . $path . ' > /dev/null 2>&1;echo "$(tail -n ' . ConfigManager::byKey('timeline::maxevent') . ' ' . $path . ')" > ' . $path);
+            ComShell::execute(SystemHelper::getCmdSudo() . 'chmod 666 ' . $path . ' > /dev/null 2>&1;echo "$(tail -n ' . ConfigManager::byKey('timeline::maxevent') . ' ' . $path . ')" > ' . $path);
             $lines = explode("\n", trim(file_get_contents($path)));
             $result = [];
             foreach ($lines as $line) {
@@ -81,7 +82,7 @@ class TimeLineHelper
     {
         $path = NEXTDOM_DATA . '/data/timeline.json';
         if (file_exists($path)) {
-            \com_shell::execute(SystemHelper::getCmdSudo() . 'chmod 666 ' . $path . ' > /dev/null 2>&1;');
+            ComShell::execute(SystemHelper::getCmdSudo() . 'chmod 666 ' . $path . ' > /dev/null 2>&1;');
             unlink($path);
         }
     }
