+++ src/Managers/NoteManager.php
@@ -38,8 +38,6 @@ use NextDom\Managers\Parents\BaseManager;
 use NextDom\Managers\Parents\CommonManager;
 use NextDom\Model\Entity\Note;
 
-require_once NEXTDOM_ROOT . '/core/class/cache.class.php';
-
 /**
  * Class NoteManager
  * @package NextDom\Managers
