+++ src/Repo/RepoMarket.php
@@ -20,6 +20,7 @@
 
 namespace NextDom\Repo;
 
+use NextDom\Com\ComShell;
 use NextDom\Enums\DateFormat;
 use NextDom\Enums\LogTarget;
 use NextDom\Exceptions\CoreException;
@@ -37,6 +38,7 @@ use NextDom\Managers\MessageManager;
 use NextDom\Managers\PluginManager;
 use NextDom\Managers\UpdateManager;
 use NextDom\Managers\UserManager;
+use NextDom\Model\Entity\JsonRPCClient;
 use NextDom\Model\Entity\Update;
 
 class RepoMarket implements BaseRepo
@@ -336,9 +338,9 @@ class RepoMarket implements BaseRepo
             if (ConfigManager::byKey('market::allowDNS') != 1 || ConfigManager::byKey('network::disableMangement') == 1) {
                 $params['url'] = NetworkHelper::getNetworkAccess('external');
             }
-            $jsonrpc = new \jsonrpcClient(ConfigManager::byKey('market::address') . '/core/api/api.php', '', $params);
+            $jsonrpc = new JsonRPCClient(ConfigManager::byKey('market::address') . '/core/api/api.php', '', $params);
         } else {
-            $jsonrpc = new \jsonrpcClient(ConfigManager::byKey('market::address') . '/core/api/api.php', '', [
+            $jsonrpc = new JsonRPCClient(ConfigManager::byKey('market::address') . '/core/api/api.php', '', [
                 'nextdomversion' => NextDomHelper::getNextdomVersion(),
                 'hwkey' => NextDomHelper::getHardwareKey(),
                 'localIp' => $internalIp,
@@ -372,7 +374,7 @@ class RepoMarket implements BaseRepo
     /**
      *
      * @param array $_arrayMarket
-     * @return \self
+     * @return self
      */
     public static function construct(array $_arrayMarket)
     {
@@ -680,7 +682,7 @@ class RepoMarket implements BaseRepo
         $cmd .= ' ' . $baseDir . '  "webdavs://' . ConfigManager::byKey('market::username') . ':' . ConfigManager::byKey('market::backupPassword');
         $cmd .= '@' . ConfigManager::byKey('market::backupServer') . '/remote.php/webdav/' . ConfigManager::byKey('market::cloud::backup::name') . '"';
         try {
-            \com_shell::execute($cmd);
+            ComShell::execute($cmd);
         } catch (\Exception $e) {
             if (self::backup_errorAnalyzed($e->getMessage()) != null) {
                 throw new CoreException('[backup cloud] ' . self::backup_errorAnalyzed($e->getMessage()));
@@ -691,7 +693,7 @@ class RepoMarket implements BaseRepo
             SystemHelper::kill('duplicity');
             shell_exec(SystemHelper::getCmdSudo() . ' rm -rf ' . $baseDir . '/tmp/duplicity*');
             shell_exec(SystemHelper::getCmdSudo() . ' rm -rf ~/.cache/duplicity/*');
-            \com_shell::execute($cmd);
+            ComShell::execute($cmd);
         }
     }
 
@@ -723,7 +725,7 @@ class RepoMarket implements BaseRepo
     {
         if (exec('which duplicity | wc -l') == 0) {
             try {
-                \com_shell::execute('sudo apt-get -y install duplicity');
+                ComShell::execute('sudo apt-get -y install duplicity');
             } catch (\Exception $e) {
 
             }
@@ -765,7 +767,7 @@ class RepoMarket implements BaseRepo
         $cmd .= ' "webdavs://' . ConfigManager::byKey('market::username') . ':' . ConfigManager::byKey('market::backupPassword');
         $cmd .= '@' . ConfigManager::byKey('market::backupServer') . '/remote.php/webdav/' . ConfigManager::byKey('market::cloud::backup::name') . '"';
         try {
-            \com_shell::execute($cmd);
+            ComShell::execute($cmd);
         } catch (\Exception $e) {
             if (self::backup_errorAnalyzed($e->getMessage()) != null) {
                 throw new CoreException('[restore cloud] ' . self::backup_errorAnalyzed($e->getMessage()));
@@ -793,10 +795,10 @@ class RepoMarket implements BaseRepo
         $cmd .= ' "webdavs://' . ConfigManager::byKey('market::username') . ':' . ConfigManager::byKey('market::backupPassword');
         $cmd .= '@' . ConfigManager::byKey('market::backupServer') . '/remote.php/webdav/' . ConfigManager::byKey('market::cloud::backup::name') . '"';
         try {
-            $results = explode("\n", \com_shell::execute($cmd));
+            $results = explode("\n", ComShell::execute($cmd));
         } catch (\Exception $e) {
             shell_exec(SystemHelper::getCmdSudo() . ' rm -rf ~/.cache/duplicity/*');
-            $results = explode("\n", \com_shell::execute($cmd));
+            $results = explode("\n", ComShell::execute($cmd));
         }
         foreach ($results as $line) {
             if (strpos($line, 'Full') === false && strpos($line, 'Incremental') === false && strpos($line, 'Complète') === false && strpos($line, 'Incrémentale') === false) {
@@ -818,7 +820,7 @@ class RepoMarket implements BaseRepo
         }
         $restore_dir = '/tmp/nextdom_cloud_restore';
         if (file_exists($restore_dir)) {
-            \com_shell::execute(SystemHelper::getCmdSudo() . ' rm -rf ' . $restore_dir);
+            ComShell::execute(SystemHelper::getCmdSudo() . ' rm -rf ' . $restore_dir);
         }
         self::backup_install();
         $base_dir = '/usr/jeedom_duplicity';
@@ -837,7 +839,7 @@ class RepoMarket implements BaseRepo
         $cmd .= '@' . ConfigManager::byKey('market::backupServer') . '/remote.php/webdav/' . ConfigManager::byKey('market::cloud::backup::name') . '"';
         $cmd .= ' ' . $restore_dir;
         try {
-            \com_shell::execute($cmd);
+            ComShell::execute($cmd);
         } catch (\Exception $e) {
             if (self::backup_errorAnalyzed($e->getMessage()) != null) {
                 throw new CoreException('[restore cloud] ' . self::backup_errorAnalyzed($e->getMessage()));
@@ -847,7 +849,7 @@ class RepoMarket implements BaseRepo
         shell_exec(SystemHelper::getCmdSudo() . ' rm -rf ' . $base_dir);
         system('cd ' . $restore_dir . ';tar cfz "' . $backup_dir . '/' . $backup_name . '" . > /dev/null');
         if (file_exists($restore_dir)) {
-            \com_shell::execute(SystemHelper::getCmdSudo() . ' rm -rf ' . $restore_dir);
+            ComShell::execute(SystemHelper::getCmdSudo() . ' rm -rf ' . $restore_dir);
         }
         BackupManager::restore($backup_dir . '/' . $backup_name, true);
     }
